service/requirements.txt
beautifulsoup4==4.9.3
blinker==1.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cos-python-sdk-v5==1.9.0
crcmod==1.7
DBUtils==2.0
dicttoxml==1.7.4
DNSPodX==0.0.1
feedparser==6.0.2
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
Pillow==8.0.1
pyasn1==0.4.8
pycryptodome==3.9.9
PyMySQL==0.10.1
PyYAML==5.3.1
qcloudsms-py==0.1.4
qrcode==6.1
requests==2.21.0
rsa==4.6
sentry-sdk==0.19.5
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.1
urllib3==1.24.3
Werkzeug==1.0.1
